requirements.txt
flask[async]
flask-cors
python-dotenv
pymongo~=4.6.2
openai-whisper
openai
whisper

--index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio

pandas
pyspark==3.2.2
